GtkIconView documentation is nearly complete.
Further to the examples directory (in which official examples used
in the manual shall be collected), I created a examples/test/ directory
in which all the internally used demo scripts are put it - the
scripts we use to find out if and how something works.
For example, the gtkiconview.phpw script has an additional window
attached to the normal iconview window. There are many buttons which
just dump out the return values of some functions or call some
functions on the iconview object